broke out sunviewsampler as own class, need to simplify sun sampler for reflections
it may be better to make a separate integrator for direct suns, as there is no overlap and everything can be recovered from
the weights array. make a sunmapper class based on view mapper, that creates a translation for each sun, so the uv coordinates can be centered
and there is less sun distortion. this class will also be needed for integration, so should be a member of the scene class